feat: functionality

+++ src/Counter.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.13;
 
 contract Counter {
-    uint256 public number;
+    uint64 public number;
 
     function setNumber(uint256 newNumber) public {
         number = newNumber;
